% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archives.r
\docType{data}
\name{archives}
\alias{archives}
\title{Archival data from tags in 2018 paper}
\format{
An object of class \code{"data.frame"}
}
\usage{
data(archives)
}
\description{
Archival data from tags in 2018 paper
}
\examples{
data(archives)
head(archives)

}
\keyword{datasets}
