% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behavioursDF.r
\docType{data}
\name{behavioursDF}
\alias{behavioursDF}
\title{Behavioural classification for the 2018 paper by Dan Fuller}
\format{
An object of class \code{"data.frame"}
}
\usage{
data(behavioursDF)
}
\description{
Behavioural classification for the 2018 paper by Dan Fuller
}
\examples{
data(behavioursDF)
head(behavioursDF)

}
\keyword{datasets}
