% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkfiles.r
\name{checkfiles}
\alias{checkfiles}
\title{This is a function to check the rdata with the archival data has been downloaded.
Git LFS is tricky}
\usage{
checkfiles()
}
\description{
This is a function to check the rdata with the archival data has been downloaded.
Git LFS is tricky
}
