% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2018.r
\docType{data}
\name{data2018}
\alias{data2018}
\title{Final dataset used in the 2018 paper. Includes behavioural classificatoins}
\format{
An object of class \code{"data.frame"}
}
\usage{
data(data2018)
}
\description{
Final dataset used in the 2018 paper. Includes behavioural classificatoins
}
\examples{
data(data2018)
head(data2018)

}
\keyword{datasets}
