% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estlength.R
\name{estlength}
\alias{estlength}
\title{This is a function to estimate length given the length at release and the time at liberty. VB Linf and K can
be provided, but default parameters for BET, based on the most recent assessment, are provided (email JH, 25/09/2024).}
\usage{
estlength(Lrel, tal, Linf = 150.314, k = 0.45104)
}
\arguments{
\item{Lrel}{Lengths at release}

\item{tal}{Times at liberty (days)}

\item{Linf}{L infinity in VB growth equation}

\item{k}{k in VB growth equation}
}
\value{
vector with lengths
}
\description{
This is a function to estimate length given the length at release and the time at liberty. VB Linf and K can
be provided, but default parameters for BET, based on the most recent assessment, are provided (email JH, 25/09/2024).
}
\examples{
#Some dummy data
dummy=data.frame(indiv=c(1,1,1,2,2,2),Lrel=c(30,30,30,80,80,80),tal=c(200,700,1500,200,700,1500))
dummy$L_rec=estlength(dummy$Lrel,dummy$tal)
dummy
}
