% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.behav.R
\name{get.behav}
\alias{get.behav}
\title{This is a function to assign an behavioural type (associated, unassociated, other)
based on the criteria from Fuller et al., 2015
https://www.sciencedirect.com/science/article/pii/S0165783615300047
Possibly many things to fine tune (also I guess the classification needs of a manual component)}
\usage{
get.behav(datetime, depth, exttemp, phase)
}
\arguments{
\item{datetime}{vector of datetime as POSIXct}

\item{depth}{vector of depths}

\item{exttemp}{vector of external temperatures}

\item{phase}{phase of day (as obtained from get.phasefromarchival or get.phasefromtrack)}
}
\value{
Returns a data.frame with the dates and behavioural classifications.
}
\description{
This is a function to assign an behavioural type (associated, unassociated, other)
based on the criteria from Fuller et al., 2015
https://www.sciencedirect.com/science/article/pii/S0165783615300047
Possibly many things to fine tune (also I guess the classification needs of a manual component)
}
\examples{
data(archives)
archivaldata=archives[archives$tagno==990289,]
archivaldata=archivaldata[complete.cases(archivaldata),]
#Only a few days to make it shorter
archivaldata$date=trunc(archivaldata$datetime,unit='days')
archivaldata=archivaldata[archivaldata$date>=unique(archivaldata$date)[26] & archivaldata$date<=unique(archivaldata$date)[35],]
data(tracks)
trackdata=tracks[tracks$tagid==990289,]
archivaldata$phase=get.phasefromtrack(archivaldata$datetime,depth=archivaldata$depth,trackdata$date,trackdata$mptlon,trackdata$mptlat,plotit=T,timearoundsunrise=c((60*60),60*60),timearoundsunset=c((60*60),60*60))
archivaldata$datetimecor=correcttime(archivaldata$datetime,mean(trackdata$mptlon))
behaviours=get.behav(archivaldata$datetimecor,archivaldata$depth,archivaldata$exttemp,archivaldata$phase)
behaviours
plot(archivaldata$datetimecor,archivaldata$depth,'p',cex=.5,ylim=c(500,0),xlab='datetime',ylab='')
for (i in 1:nrow(behaviours)){
  behavcolor=ifelse(behaviours$behav[i]=='Associated','red',
                    ifelse(behaviours$behav[i]=='Unassociated','blue','grey'))
  rect(as.numeric(behaviours$date[i]),500,as.numeric(behaviours$date[i])+86400,480,col=behavcolor)
}
legend('bottomright',legend = c("Associated", "Unassociated","Other"), 
       col = c("red", "blue","grey"), pch = 15, pt.cex = 2, 
       cex = 1.2, title = "Behavioural type")
}
