% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.phasefromtrack.R
\name{get.phasefromtrack}
\alias{get.phasefromtrack}
\title{This is a function to split archival data between dawn, day, dusk and night.
based on the geolocations, estimating sunrise and sunset times and defining
periods around sunrise and sunset.
datetime,depth,geodates,geolons,geolats}
\usage{
get.phasefromtrack(
  datetime,
  depth,
  geodates,
  geolons,
  geolats,
  plotit = F,
  timearoundsunrise = c((30 * 60), 0),
  timearoundsunset = c((30 * 60), 0)
)
}
\arguments{
\item{datetime}{vector of datetime as POSIXct}

\item{depth}{vector of depths (numeric)}

\item{geodates}{vector of dates from the geolocation file}

\item{geolons}{vector of longitudes from the geolocation file}

\item{geolats}{vector of latitudes from the geolocation file}

\item{plotit}{plot example (default to F)}

\item{timearoundsunrise}{numeric vector with two values defining the number of seconds before and after sunrise the define the dawn}

\item{timearoundsunset}{numeric vector with two values defining the number of seconds before and after sunrise the define the dawn}
}
\value{
Returns a character vector with phases of the day for each record.
}
\description{
This is a function to split archival data between dawn, day, dusk and night.
based on the geolocations, estimating sunrise and sunset times and defining
periods around sunrise and sunset.
datetime,depth,geodates,geolons,geolats
}
\examples{
data(archives)
archivaldata=archives[archives$tagno==990289,]
archivaldata=archivaldata[complete.cases(archivaldata),]
data(tracks)
trackdata=tracks[tracks$tagid==990289,]
archivaldata$phase=get.phasefromtrack(archivaldata$datetime,archivaldata$depth,trackdata$date,trackdata$mptlon,trackdata$mptlat,plotit=T,timearoundsunrise=c((60*60),60*60),timearoundsunset=c((60*60),60*60))
archivaldata$date=trunc(archivaldata$datetime,unit='days')
seldate=unique(archivaldata$date)[31]
plot(archivaldata$datetime[archivaldata$date==seldate],archivaldata$depth[archivaldata$date==seldate],'p',pch=16,xlab='',ylab='',ylim=c(300,0))
points(archivaldata$datetime[archivaldata$date==seldate & archivaldata$phase=='day'],archivaldata$depth[archivaldata$date==seldate & archivaldata$phase=='day'],pch=16,col='red')
}
