% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.thermalstruct.R
\name{get.thermalstruct}
\alias{get.thermalstruct}
\title{This is a function to estimate the depth of different isoclines for each day from an archival tagging dataset}
\usage{
get.thermalstruct(datetime, depth, exttemp, isoclines = c(18, 20))
}
\arguments{
\item{datetime}{vector of datetime as POSIXct}

\item{depth}{vector of depths (numeric)}

\item{exttemp}{vector of temperatures (numeric)}

\item{isoclines}{vector of isoclines (default c(18,20))}
}
\value{
Returns a data.frame with the depth of the different isoclines by day
}
\description{
This is a function to estimate the depth of different isoclines for each day from an archival tagging dataset
}
\examples{
data(archives)
archivaldata=archives[archives$tagno==990289,]
archivaldata=archivaldata[complete.cases(archivaldata),]
archivaldata$date=trunc(archivaldata$datetime,unit='days')
archivaldata=archivaldata[archivaldata$date<unique(archivaldata$date)[10],]
res=get.thermalstruct(archivaldata$datetime,archivaldata$depth,archivaldata$exttemp, isoclines=c(18,20,22))
res
}
