% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getoceano.R
\name{getoceano}
\alias{getoceano}
\title{Function to interpolate values of oceanographic data (mixed layer depth or depth of an isocline)}
\usage{
getoceano(
  product = c("mld", "isoclinedepth"),
  foldername,
  lons,
  lats,
  years,
  months,
  isocline = NULL
)
}
\arguments{
\item{product}{Product to interpolate Mixed layer depth (mld) or Potential temperature (temp)}

\item{foldername}{folder with data}

\item{lons}{longitudes in decimal degrees}

\item{lats}{latitudes in decimal degrees}

\item{years}{vector with years}

\item{months}{vector with months}

\item{isocline}{isocline to interpolate (Celsius)}
}
\value{
vector with interpolated depth values
}
\description{
Function to interpolate values of oceanographic data (mixed layer depth or depth of an isocline)
}
\examples{
dir.create('oceanodata')
downloadgodas(product='temp',year=1999,minlon=120,maxlon=240,minlat=-40,maxlat=40,maxdepth=500,outputfolder = 'oceanodata')
isodepth('oceanodata',year=1999,isocline=20)
getoceano(product = 'isoclinedepth',foldername='oceanodata',lons=c(145,164,192),lats=c(-15,-3.5,24),years=c(1999,1999,1999),months=c(1,3,3),isocline=20)
}
