% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isodepth.R
\name{isodepth}
\alias{isodepth}
\title{Function to read files of potential temperatures from godas, estimate the depth of an isocline and save it as an RDS object
in the same folder}
\usage{
isodepth(foldername, year, isocline)
}
\arguments{
\item{foldername}{folder with temperature data}

\item{year}{year (numeric)}

\item{isocline}{isocline depth to be estimated (Celsius)}

\item{product}{Product to download. Mixed layer depth (mld) or Potential temperature (temp)}
}
\value{
Saves the download as an RDS file in the output folder
}
\description{
Function to read files of potential temperatures from godas, estimate the depth of an isocline and save it as an RDS object
in the same folder
}
\examples{
dir.create('oceanodata')
downloadgodas(product='temp',year=1999,minlon=120,maxlon=240,minlat=-40,maxlat=40,maxdepth=500,outputfolder = 'oceanodata')
isodepth('oceanodata',year=1999,isocline=20)
dat=readRDS('./oceanodata/depth_20_isocline_1999.rds')
require(raster)
dat2=raster(t(dat$depth[,dim(dat$depth)[2]:1,1]))
crs(dat2) <- "+proj=longlat +datum=WGS84"
extent(dat2) <- c(range(dat$lons),range(dat$lats))
x11(10,10)
plot(dat2,main='depth 20 degs isocline in jan 1999')
}
