% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md2018.r
\docType{data}
\name{md2018}
\alias{md2018}
\title{Metadata tags in 2018 paper}
\format{
An object of class \code{"data.frame"}
}
\usage{
data(md2018)
}
\description{
Metadata tags in 2018 paper
}
\examples{
data(md2018)
head(md2018)

}
\keyword{datasets}
