% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermstruct.r
\docType{data}
\name{thermstruct}
\alias{thermstruct}
\title{Isocline depth by tag and days estimated from the archival data (function get.thermalstruct)}
\format{
An object of class \code{"data.frame"}
}
\usage{
data(thermstruct)
}
\description{
Isocline depth by tag and days estimated from the archival data (function get.thermalstruct)
}
\examples{
data(thermstruct)
head(thermstruct)

}
\keyword{datasets}
