% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracks.r
\docType{data}
\name{tracks}
\alias{tracks}
\title{Tracks of tags in 2018 paper}
\format{
An object of class \code{"data.frame"}
}
\usage{
data(tracks)
}
\description{
Tracks of tags in 2018 paper
}
\examples{
data(tracks)
head(tracks)

}
\keyword{datasets}
