% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkfiles.r
\name{checkfiles}
\alias{checkfiles}
\title{This is a function to check the rdata with the archival data has been downloaded.
After some issues with LFS, I put it in saco...}
\usage{
checkfiles()
}
\description{
This is a function to check the rdata with the archival data has been downloaded.
After some issues with LFS, I put it in saco...
}
